requirements.txt
numpy>=1.26.4
ocrmac>=1.0.1; platform_system == "Darwin"
openai>=1.51.0
pydantic>=2.9.0
pymysql>=1.1.1
pymupdf>=1.24.10
pypdf>=5.1.0
pytesseract>=0.3.13
rapidocr-onnxruntime>=1.2.3; python_version < "3.13"
python-dotenv>=1.0.1
requests>=2.32.3
pytest>=8.3.0
